refactor(web): split MCP management page flows
